Fixed second write issue, introduced by PR#3675
